requirements.txt
flask
selenium
openai
google-generativeai
webdriver-manager
gunicorn
